pyproject.toml
[tool.poetry]
name = "yaml-config"
version = "0.1.0"
description = "Manage YAML based configs"
authors = ["Ids van der Molen <2285220+idsvandermolen@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
kubernetes = "^24.2.0"
pylint = "^2.14.2"
pytest = "^7.1.3"
datapath = {git = "https://github.com/idsvandermolen/datapath.git", rev = "v0.1.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
